Updating on_create.sh to install and configure EFA client for FSx. Making it consistent to Slurm, which now supports EFA for fsx.
The change is checking and making sure the OS is supported for EFA backed FSx and the instance has EFA available before proceeding with client installation.

Since fsx is mounted later with eks, we can't verify if fsx is efa enabled or not beforehand, or if the fsx and instance are in the same AZ, but in that case, fsx will automatically fall back to use TCP instead of EFA, so, there is no drawback in installing the client at provisioning time.